fix(serial): P2 Bug修复x6+连接健康UI+safeFromUtf8+DataStatistics会话摘要+QSS三主题框架样式#58
P2-01: TerminalWidget clear()同步清空m_model防止旧数据闪现
P2-02: safeFromUtf8匿名命名空间辅助函数+Text/Mixed模式替换fromUtf8(无效字节→\xHH)
P2-05: ToastWidget字体fallback链(跨平台: Segoe UI Emoji→Noto Color Emoji→Apple Color Emoji)
P2-07: OtaWidget animateCompletion QPropertyAnimation误用→QTimer::singleShot(400ms)
P2-UB: FrameDefinition UInt32LE/BE unsigned char→uint32_t移位提升修复
P2-Signal: RecordingController blockSignals防setChecked递归触发onToggleRecording
连接健康: connectionHealth信号→DataStatistics空闲提示(>10s显示)QSS三主题healthStatusLabel
重连UI: reconnectProgress信号→Toast防抖通知(含指数退避倒计时)
DataStatistics: sessionSummary()会话导出摘要+totalRx/totalTx访问器+QFrame分组样式(statsRx/Tx/Peak/Time)
QSS: 三主题新增DataStatistics框架样式(RX绿/TX红/峰值蓝/时间灰 半透明卡片)

评分: 58 + 1 = 59
变更: 13 files, +342 insertions, -53 deletions

+++ src/protocol/FrameDefinition.cpp
@@ -64,18 +64,19 @@ QVariant FieldDef::extractValue(const QByteArray& payload) const
             (static_cast<unsigned char>(d[1])));
         break;
     case UInt32LE:
+        // 先转 uint32_t 再移位，避免 unsigned char→int 提升后 << 24 越界（UB）
         raw = static_cast<uint32_t>(
-            (static_cast<unsigned char>(d[0])) |
-            (static_cast<unsigned char>(d[1]) << 8) |
-            (static_cast<unsigned char>(d[2]) << 16) |
-            (static_cast<unsigned char>(d[3]) << 24));
+            (static_cast<uint32_t>(static_cast<unsigned char>(d[0]))) |
+            (static_cast<uint32_t>(static_cast<unsigned char>(d[1])) << 8) |
+            (static_cast<uint32_t>(static_cast<unsigned char>(d[2])) << 16) |
+            (static_cast<uint32_t>(static_cast<unsigned char>(d[3])) << 24));
         break;
     case UInt32BE:
         raw = static_cast<uint32_t>(
-            (static_cast<unsigned char>(d[0]) << 24) |
-            (static_cast<unsigned char>(d[1]) << 16) |
-            (static_cast<unsigned char>(d[2]) << 8) |
-            (static_cast<unsigned char>(d[3])));
+            (static_cast<uint32_t>(static_cast<unsigned char>(d[0])) << 24) |
+            (static_cast<uint32_t>(static_cast<unsigned char>(d[1])) << 16) |
+            (static_cast<uint32_t>(static_cast<unsigned char>(d[2])) << 8) |
+            (static_cast<uint32_t>(static_cast<unsigned char>(d[3]))));
         break;
     case Int8:
         raw = static_cast<signed char>(d[0]);
